1. Rolled back FatFs to 0.04b. The mmc.c driver in use was originally developed by LMI for 0.04b and does not implement some functionality in FatFs 0.06. Although it worked fine on my test system, there could be potential problems with some boards and/or cards.
2. eLua currently does not switch the JTAG pins to GPIO. This causes
some GPIO pins (PB7 and PC0-3 on the LM3S6918) to be non-functional for
eLua applications. Added jtag.c and jtag.h to src/platform/lm3s to
implement a workaround. Will check with the mailing list for the proper
way to handle this within eLua before changing the base code.