Fixed bug where re-use of interrupt vector configuration struct caused UART interrupt disable.
As it stands, this should now work pretty well as long as buffers are in place
on the target.  Perhaps some additional synchronous checks should be added so
that the server is expected to acknowledge the command type (indicating that
pending operations are clear, and it will have enough captive polling ability
to avoid losing chars).  This could also be useful for situations where one
end may support more commands than the other, preventing a client from
spitting out a full command if the server doesn't support it?