Refactored version of TetrIves game (for the LM3S.disp issue). Minor remarks on Pong & SpaceShip games.
Tests on both LM3S platforms that have a RIT OLED display are welcome.
An LM3S6965.lua file must be created and commited, so they can work on this board.
A simle look at the LM3S8962.lua will be enough to understand what needs to be adjusted for the pin assignments.
This scheme leaves the programs more portable, with a simple require(pd.cpu()) to "include" pinout definitions.
Pls note that in the future this may change to require(pd.board()) but for now we have a 14 char limitation on our ROMFS filenames and we don't want to postpone v0.6 anymore (comming soon now ! :)